Got user sign in working again
Got user sign in working again in this patch. User model had a
remember_me! method which had save_without_validation commented
because it was not available in rails 3. This patch replaces the
remember_me! method by having an expectation that the remember_me
token will generated when the user is first saved in the database

Modified tests to work around this new expectation